requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
google-generativeai
python-dotenv
playwright
selenium
beautifulsoup4
pandas
webdriver-manager